+++ wgpu-core/src/device/resource.rs
@@ -4163,9 +4163,6 @@ impl Device {
             });
         }
 
-        if desc.immediate_size != 0 {
-            self.require_features(wgt::Features::IMMEDIATES)?;
-        }
         if self.limits.max_immediate_size < desc.immediate_size {
             return Err(Error::ImmediateRangeTooLarge {
                 size: desc.immediate_size,