+++ wgpu-core/src/indirect_validation/dispatch.rs
@@ -93,7 +93,7 @@ impl Dispatch {
         let module = panic!("Indirect validation requires the wgsl feature flag to be enabled!");
 
         let info = crate::device::create_validator(
-            wgt::Features::IMMEDIATES,
+            wgt::Features::empty(),
             wgt::DownlevelFlags::empty(),
             naga::valid::ValidationFlags::all(),
         )