+++ wgpu-core/src/timestamp_normalization/mod.rs
@@ -151,7 +151,7 @@ impl TimestampNormalizer {
                 panic!("Timestamp normalization requires the wgsl feature flag to be enabled!");
 
             let info = crate::device::create_validator(
-                wgt::Features::IMMEDIATES,
+                wgt::Features::empty(),
                 wgt::DownlevelFlags::empty(),
                 naga::valid::ValidationFlags::all(),
             )