+++ wgpu/src/api/pipeline_layout.rs
@@ -39,8 +39,6 @@ pub struct PipelineLayoutDescriptor<'a> {
     /// The number of bytes of immediate data that are allocated for use
     /// in the shader. The `var<immediate>`s in the shader attached to
     /// this pipeline must be equal or smaller than this size.
-    ///
-    /// If this value is non-zero, [`Features::IMMEDIATES`] must be enabled.
     pub immediate_size: u32,
 }
 #[cfg(send_sync)]