+++ wgpu/src/backend/webgpu/webgpu_sys/gen_GpuComputePassEncoder.rs
@@ -338,7 +338,7 @@ extern "C" {
     pub fn set_immediates_with_u8_slice(
         this: &GpuComputePassEncoder,
         range_offset: u32,
-        data: &mut [u8],
+        data: &[u8],
     ) -> Result<(), JsValue>;
 
     #[wasm_bindgen(